Use testcontainers instead of H2
### What's done:
* Use shared DB container using base class DatabaseTestBase.kt
